+++ app/Http/Controllers/TagsController.php
@@ -99,6 +99,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        // If this tag is associated with any post, Don't allow the deletion
+        if($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted, because it is associated with some posts');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
